Store Book data in more retrievable format
The book data was stored in a vector, which does not allow efficient data lookups. In order to allow for better retrieval the data is stored in a map using the book code, which is unique for each book, as an index.
This allows other data files to use this as a primary key for reference rather than the book name, which contains spaces and other details making it a poor choice for primary key.